Add `ReadOnlyGraph` interface
The `ReadOnlyGraph` is an interface for a graph that cannot be mutated.
As such, it only presents its read methods, e.g. `nodes`.

This change is motivated by my desire to create "enriched" graph
classes, e.g. `WeightedGraph`, which wrap the graph along with
additional metadata. For those enriched graphs to maintain integrity,
independently modifying the underlying graph must be disallowed, so this
is the interface they should implement.

In principle, I'd like to add `toJSON` and `equals` to this interface,
but it's not clear how to do that properly.

Test plan: `yarn flow` suffices.