Add the `WeightedGraph` class
As discussed in depth in #1004, it makes sense to package a Graph along
with the metadata needed to run Pagerank on it: hence, the
WeightedGraph, which is basically a Graph bundled with weights for every
edge, and a synthetic self loop weight.

In #1008, I created WeightedGraph as a simple struct. This commit
explores a different approach: WeightedGraph is reified as a full class
(a sibling to Graph) with a constructor that validates its invariants,
JSON (de)serialization, equality testing, and functions that implement
basic behaviors on top of it. The potential benefits of this approach
are discussed in
[this comment](https://github.com/sourcecred/sourcecred/issues/1004#issuecomment-439641568)

The intention is that consumers that need read-only access to a Graph
should be able to use a WeightedGraph as well. To that end, the
WeightedGraph implements the ReadOnlyGraph interface.

One possible improvement: we could add `addNode` and `addEdge` to the
WeightedGraph, but with semantics that weights must be present. So,
`addEdge` requires an edge and a weight to both be provided. It wouldn't
be hard to implement. This is not needed by any existing client, but
having it could make it much easier to directly create PageRank-able
graphs. (If you don't need the decoupling offered by providing an edge
evaulator separately from specifiying the graph, it is really annoying.
I learned this while trying to create WeightedGraph test cases.)

Test plan: New, thorough unit tests have been added. Run `yarn test`.