Fix markov chain semantics for empty graphs
While working on #1020, I discovered that attempting to find the
stationary distribution on an empty graph currently throws an error.
This is awkward, since the empty graph is a totally reasonable graph for
our code to work on.

This commit modifies the markov chain semantics so that we can consider
empty chains (which have no probability mass). Attempting to find the
stationary distribution of one immediately returns an empty
distribution, after 0 iteration steps and with a convergence delta of
`NaN`.

Test plan: Unit tests included; `yarn test` should suffice.