schema: annotate node fields with `Fidelity`
This commit addresses (#998) and (#996) and corresponds to
1.) in @wchargin's suggested implementation plan.

We equip each node field with a _fidelity_: a field that is
known to sometimes return results of incorrect type is “unfaithful”, and
is accompanied by a list of typenames that can be returned in
practice; other fields are “faithful”, which has implicitly been
the case for all fields until now.

Existing consumers of schemata (the Mirror module and Flow type
generation) do not yet support unfaithful fields.

Test Plan:

Added tests to `schema.test.js` to check validity of Unfaithful
Fidelity and duplicate `actualTypenames`.

Added tests to check the enforcement of `Not Yet Implemented`
errors.