handles `TypeNameUpdateResults` in `_queryFromPlan`
The following handles `typenames` in `_queryFromPlan`.

This commit addresses (#998) and (#996) and corresponds to
6.) in @wchargin's suggested implementation plan.

Typenames are first paginated by the total `nodesLimit`.
They are then paginated with an explicit `typnamesLimit`
number, which has been added to options in all previous
QueryPlans (including tests).

A GraphQL query is then made to fetch the `id`,
`__typename`, and top level primitive fields
 of all the ids in the `typenames` list.

Test Plan: A new test was added to `mirror.test.js`
which checks for the appropriate pagination of
`typenamesLimit`.

Previous tests were edited to include the `typenamesLimit`
in options.

All new and subsequent tests pass.