updates `_updateData` to handle a `TYPENAME_UPDATE`
This commit addresses (#998) and (#996) and corresponds to
7.) and 8.) in @wchargin's suggested implementation plan.

Updates `_updateData` to handle a `TYPENAME_UPDATE`.
For updates containing data corresponding to an
unfaithful typename,
`_nontransactionallyRegisterTypename` will be called
if the object has not been previously registered in
the db. This will insert the object in the `objects`
table with a `null` `typename` and `null` `last_update`

If the object has been previously registered,
`_nontransactionallyUpdateTypename` will be called,
setting the appropriate canonical `typename`, as well
as the top level primitives for the object.

If an object is found, `last_update` and `typename` are
updated with the corresponding `updateRecord` from the
`TypenameUpdateResult`.

Test Plan:
Added tests `_getTypename` as a helper function to
retrieve the typename of an object in the database from
its `id`.

Added tests for `_updateTypename`, `_registerTypename`,
and querying an unfaithful typename in `mirror.update()`