[draft] Add an org-mode to `sourcecred load`
It's common for a project to be split up under many repositories; e.g.
ipfs has 238. We have a feature that allows the user to aggregate cred
for many repositories into one output, but doing so for 238 repos is
incredibly tedious.

This commit adds a `--organization` flag to `sourcecred load` which
loads all non-fork repositories for a given organization, and saves it
to the repoId `organization/combined`.

The list of non-fork repositories is retrieved by using the GitHub
GraphQL search API. Currently, this is implemented without the benefit
of re-using any of our other GraphQL logic; we should probably
de-duplicate it before merging.

Test plan: I've used this for several orgs for demos and it's worked;
hoewver, we should test it before merging.