adds typenames field to `QueryPlan`
This commit addresses (#998) and (#996) and corresponds to
2.) in @wchargin's suggested implementation plan.

This adds a `typenames` field to the `QueryPlan` type
in Mirror. `typenames` is a list of GraphQL id's to be queried for
their canonical typename.

`_findOutdated` in Mirror has been updated to an empty list for
this new field.

`_queryFromPlan` has been updated to throw a `Not Yet Implemented`
Error if `queryPlan.typenames` is not empty.

Test Plan: Tests in `mirror.test.js` have been updated to include
an list for a `typenames` field anywhere a `QueryPlan` is.

A test has been added to `_queryFromPlan` that checks that
`typenames` will throw a NYE error given a non-null list.