Graph Visualizer: Add NodeRenderer
This commit adds a `NodeRenderer` class for visualizing nodes in the
SourceCred graph. The NodeRenderer is a react component which maintains
a d3 selection under the hood. This approach allows us to build the
visualizer around a React architecture, while still benefiting from d3
transitions. See [this blog post] for details.

[this blog post]: https://medium.com/@sxywu/on-d3-react-and-a-little-bit-of-flux-88a226f328f3

Rather than writing unit tests, I wrote an 'Inspection Test',
which is basically a visual component which drives all the features of
the NodeRenderer, and allows manual inspection.

As compared to a unit test, it has several advantages:
- Allows testing transitions
- Easier/faster to write
- Since the output is visual, visual inspection can be better at finding
bugs than unit testing

Its main disadvantage is that since it's a manual test, people might
forget to run it, and inadvertently break the application. However,
since the visualizer is quite self-contained (few dependencies into the
rest of the application), I don't think this is a big concern.

Test plan: Run the inspection test (by running `yarn start` and then
navigating to `/test/NodeRenderer/`), and manually verify the behavior
is correct.